pyproject.toml
[tool.poetry]
name = "mygitproj"
version = "0.1.0"
description = ""
authors = ["AshDavid12 <gaashdavid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
langchain = "^0.2.7"
langchain-openai = "^0.1.14"
langchain-core = "^0.2.12"
youtube-transcript-api = "^0.6.2"
pytube = "^15.0.0"
streamlit = "^1.36.0"
pydantic = "^2.8.2"
groq = "^0.9.0"
langchain-groq = "^0.1.6"
git-filter-repo = "^2.38.0"
python-dotenv = "^1.0.1"
click = "^8.1.7"
gitpython = "^3.1.43"
openai = "^1.35.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
